separate main code
