+++ src/mapOptmization.cpp
@@ -1497,50 +1497,4 @@ bool mapOptimization::LMOptimization(int iterCount)
     return false; // keep optimizing
 }
 
-int main(int argc, char** argv)
-{
-    rclcpp::init(argc, argv);
-
-    rclcpp::NodeOptions options;
-    options.use_intra_process_comms(true);
-    rclcpp::executors::SingleThreadedExecutor exec;
-
-    auto MO = std::make_shared<mapOptimization>(options);
-    exec.add_node(MO);
-
-    // Loop Closure 스레드 시작
-    std::thread loopThread;
-    if (MO->loop_closure_) {
-        try {
-            RCLCPP_INFO(MO->get_logger(), "Starting Loop Closure thread...");
-            loopThread = std::thread(&LoopClosure::loopClosureThread, MO->loop_closure_.get());
-        } catch (const std::exception& e) {
-            RCLCPP_ERROR(MO->get_logger(), "Failed to start Loop Closure thread: %s", e.what());
-        }
-    } else {
-        RCLCPP_WARN(MO->get_logger(), "Loop Closure module not initialized");
-    }
-
-    std::thread visualizeMapThread(&mapOptimization::visualizeGlobalMapThread, MO);
-    
-    // 코스트맵 스레드는 CostmapGenerator 내부에서 시작
-    if (MO->costmap_generator_) {
-        MO->costmap_generator_->startCostmapThread();
-    }
-
-    exec.spin();
-
-    rclcpp::shutdown();
-
-    // 안전하게 스레드 종료
-    if (loopThread.joinable()) {
-        loopThread.join();
-    }
-    
-    if (visualizeMapThread.joinable()) {
-        visualizeMapThread.join();
-    }
-    // costmapThread는 ~CostmapGenerator()에서 자동으로 정리됨
-
-    return 0;
-}
+// main 함수는 별도 파일(src/main.cpp)로 분리되었습니다.
